+++ packages/fighting-design/novel-coronavirus/index.ts
@@ -1,9 +1,0 @@
-import FNovelCoronavirus from './src/novel-coronavirus.vue'
-
-import { install } from '../_utils'
-
-install(FNovelCoronavirus, FNovelCoronavirus.name)
-
-export type FNovelCoronavirusInstance = InstanceType<typeof FNovelCoronavirus>
-
-export default FNovelCoronavirus