+++ start/src/App.vue
@@ -1,32 +1,17 @@
 <script lang="ts" setup></script>
 
 <template>
-  <div class="action">
-    <f-button type="primary" size="large" href="/docs/install">
-      开始使用
-    </f-button>
-    <f-button
-      class="code"
-      type="primary"
-      size="large"
-      blob
-      simple
-      data-clipboard-text="npm i filling-design"
-    >
-      pnpm add filling-design
-    </f-button>
-  </div>
+  <!-- <f-image
+    width="200px"
+    src="https://abc.com/1.png"
+    err-src="https://tianyuhao.cn/images/auto/1.jpg"
+  /> -->
+  <f-image
+    width="200px"
+    src="https://abc.com/1.png"
+    err-src="https://abc.com/1.png"
+    alt="error"
+  />
 </template>
 
-<style lang="scss" scoped>
-  .action {
-    margin-top: 30px;
-    display: flex;
-    align-items: center;
-
-    .code {
-      margin-left: 17px;
-      font-size: 21px;
-    }
-  }
-</style>
+<style lang="scss" scoped></style>